Protect download routes
